+++ captura/views.py
@@ -300,9 +300,13 @@ def estudio_delete(request):
         return redirect('captura:estudios')
     return HttpResponseBadRequest()
 
+
 @login_required
 @user_passes_test(is_capturista)
 def recover_studies(request):
+    """ View to list the studies that are deleted and can be recovered.
+
+    """
     return render(request, 'captura/recuperar_estudios.html')
